more

+++ db/migrate/20140909062020_create_data_rows.rb
@@ -3,13 +3,13 @@ class CreateDataRows < ActiveRecord::Migration
     create_table :data_rows do |t|
       t.integer :sub_data_table_id
       t.integer :year
-      t.double :t
-      t.double :i
-      t.double :small_n
-      t.double :big_n
-      t.double :p
-      t.double :r
-      t.double :d
+      t.decimal :t
+      t.decimal :i
+      t.decimal :small_n
+      t.decimal :big_n
+      t.decimal :p
+      t.decimal :r
+      t.decimal :d
 
       t.timestamps
     end
